// packages/plugin-freshdesk/src/types/FreshDeskTicket.ts
export type FreshDeskTicket = {
  cc_emails: string[];
  fwd_emails: string[];
  reply_cc_emails: string[];
  ticket_cc_emails: string[];
  fr_escalated: boolean;
  spam: boolean;
  email_config_id: number;
  group_id: number | null;
  priority: number;
  requester_id: number;
  responder_id: number;
  source: number;
  company_id: number;
  status: number;
  subject: string;
  association_type: number | null;
  support_email: string;
  to_emails: string[];
  product_id: number;
  id: number;
  type: string;
  due_by: string;
  fr_due_by: string;
  is_escalated: boolean;
  description: string;
  description_text: string;
  custom_fields: Record<string, unknown>;
  created_at: string;
  updated_at: string;
  tags: string[];
  attachments: unknown[];
  source_additional_info: unknown | null;
};